Cluster aes session rotation test
